requirements.txt
torch
torchvision
torchaudio
ultralytics~=8.2.1
opencv-python~=4.9.0.80
matplotlib~=3.8.4
pillow~=10.3.0
numpy~=1.26.4
diffusers~=0.27.2
fastapi
uvicorn